+++ src/iomei.cpp
@@ -58,6 +58,7 @@
 #include "octave.h"
 #include "page.h"
 #include "pages.h"
+#include "pb.h"
 #include "pedal.h"
 #include "pgfoot.h"
 #include "pgfoot2.h"
@@ -68,6 +69,7 @@
 #include "rend.h"
 #include "rest.h"
 #include "rpt.h"
+#include "sb.h"
 #include "score.h"
 #include "section.h"
 #include "slur.h"
