+++ src/iomusxml.cpp
@@ -37,10 +37,12 @@
 #include "mrest.h"
 #include "note.h"
 #include "octave.h"
+#include "pb.h"
 #include "pedal.h"
 #include "rend.h"
 #include "rest.h"
 #include "rpt.h"
+#include "sb.h"
 #include "score.h"
 #include "section.h"
 #include "slur.h"
