+++ src/section.cpp
@@ -71,44 +71,6 @@ void Section::AddChild(Object *child)
 }
 
 //----------------------------------------------------------------------------
-// Pb
-//----------------------------------------------------------------------------
-
-Pb::Pb() : SystemElement("pb-"), AttNNumberLike()
-{
-    RegisterAttClass(ATT_NNUMBERLIKE);
-
-    Reset();
-}
-
-Pb::~Pb() {}
-
-void Pb::Reset()
-{
-    SystemElement::Reset();
-    ResetNNumberLike();
-}
-
-//----------------------------------------------------------------------------
-// Sb
-//----------------------------------------------------------------------------
-
-Sb::Sb() : SystemElement("sb-"), AttNNumberLike()
-{
-    RegisterAttClass(ATT_NNUMBERLIKE);
-
-    Reset();
-}
-
-Sb::~Sb() {}
-
-void Sb::Reset()
-{
-    SystemElement::Reset();
-    ResetNNumberLike();
-}
-
-//----------------------------------------------------------------------------
 // Section functor methods
 //----------------------------------------------------------------------------
 
@@ -161,48 +123,6 @@ int Section::ResetDrawing(FunctorParams *functorParams)
     }
 
     return FUNCTOR_CONTINUE;
-}
-
-//----------------------------------------------------------------------------
-// Pb functor methods
-//----------------------------------------------------------------------------
-
-int Pb::CastOffEncoding(FunctorParams *functorParams)
-{
-    CastOffEncodingParams *params = dynamic_cast<CastOffEncodingParams *>(functorParams);
-    assert(params);
-
-    if (!params->m_firstPbProcessed) {
-        params->m_firstPbProcessed = true;
-    }
-    else {
-        params->m_currentPage = new Page();
-        assert(params->m_doc->GetPages());
-        params->m_doc->GetPages()->AddChild(params->m_currentPage);
-        params->m_currentSystem = new System();
-        params->m_currentPage->AddChild(params->m_currentSystem);
-    }
-
-    MoveItselfTo(params->m_currentSystem);
-
-    return FUNCTOR_SIBLINGS;
-}
-
-//----------------------------------------------------------------------------
-// Sb functor methods
-//----------------------------------------------------------------------------
-
-int Sb::CastOffEncoding(FunctorParams *functorParams)
-{
-    CastOffEncodingParams *params = dynamic_cast<CastOffEncodingParams *>(functorParams);
-    assert(params);
-
-    params->m_currentSystem = new System();
-    params->m_currentPage->AddChild(params->m_currentSystem);
-
-    MoveItselfTo(params->m_currentSystem);
-
-    return FUNCTOR_SIBLINGS;
 }
 
 } // namespace vrv
